requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0.0
asyncpg
pydantic-settings
python-multipart
gunicorn
